// frontend/src/api/types/TemplateDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TagRef } from "./TagRef";
import type { TopicRef } from "./TopicRef";

export type TemplateDetail = { id: string, name: string, description: string, title: string, body_md: string, topics: Array<TopicRef>, tags: Array<TagRef>, updated_at: string, updated_by_name: string, };